tls/requirements-update-tls.d6.txt
# requirements for update-tls.d6.py

cloudera-director-python-client >= 6.0.0
